Use a buffer pool for read response data
The pool uses 32 * 4 KiB buffers, which should match the maximum size of
a read request on all fuse implementation. Linux with FUSE protocol 7.28
support up to 256 pages, but the fuse package currently does not support
it and even after patching it in, reads using dd with direct io and a
large enough block size where never larger than 128 KiB, so it should be
safe to hardcode 32 pages for now.