Fix reads across chunk boundaries
When using Direct-IO requests are not guaranteed to be aligned on page
size boundary, like with buffered IO. This means a read request can
start at the end of one chunk and finish at the beginning of the next
chunk. In cases where this happens we now request multiple chunks in
GetChunk and assemble the response data.