Remove extra buffers for load threads
This makes it easier to calculate the size of the buffer pool, since it
is directly derived from max-chunks, instead of adding additional
buffers for each load thread.

However this seems to somewhat increase lock contention, since getting a
download buffer now has to go through the storage and its two mutexes to
reserve capacity and drop old buffers if none are available.